Move around testing-only code.
